move keypress stuff to process_keypress.c
