+++ src/Makefile
@@ -1,9 +1,9 @@
 all: kilo
 
-SRC=kilo.c function.c
+OBJS=kilo.o function.o process_keypress.o
+
 INCLUDE=function.h kilo.h
 CFLAGS=-Wextra -pedantic -std=c11
-OBJS=kilo.o function.o
 
 kilo: $(OBJS)
 	$(CC) -o kilo $(OBJS)
