+++ src/process_keypress.h
@@ -1,0 +1,6 @@
+#ifndef KILO_PROCESS_KEYPRESS_H
+#define KILO_PROCESS_KEYPRESS_H
+
+void editorProcessKeypress(int fd);
+
+#endif
